// src/infrastructure/infrastructure.provider.enum.ts
export enum InfrastructureProviderEnum {
  REDIS_SERVICE = 'INFRASTRUCTURE_REDIS_SERVICE',
  LOGGER_SERVICE = 'INFRASTRUCTURE_LOGGER_SERVICE',
  HTTP_SERVICE = 'INFRASTRUCTURE_HTTP_SERVICE',
  EXAMPLE_HTTP_SERVICE = 'INFRASTRUCTURE_EXAMPLE_HTTP_SERVICE',

  // Postgres
  POSTGRES_DATASOURCE = 'INFRASTRUCTURE_POSTGRES_DATASOURCE',
  PROFILE_REPOSITORY = 'INFRASTRUCTURE_PROFILE_REPOSITORY'
}