// src/domain/domain.provider.enum.ts
export enum DomainProviderEnum {
  HEALTH_CHECK_USE_CASE = 'DOMAIN_HEALTH_CHECK_USE_CASE',
  CREATE_PROFILE_USE_CASE = 'DOMAIN_CREATE_PROFILE_USE_CASE',
  CREATE_PROFILE__FROM_XLS_USE_CASE = 'DOMAIN_CREATE_PROFILE_FROM_XLS_USE_CASE',
  GET_PROFILE_BY_EMAIL_USE_CASE = 'DOMAIN_GET_PROFILE_BY_EMAIL_USE_CASE',
  DELETE_PROFILE_BY_ID_USE_CASE = 'DOMAIN_DELETE_PROFILE_BY_ID_USE_CASE',
  UPDATE_PROFILE_USE_CASE = 'DOMAIN_UPDATE_PROFILE_USE_CASE',
}